+++ .travis.yml
@@ -1,11 +1,33 @@
 sudo: required
 dist: trusty
 
-install:
-  - sudo apt-add-repository -y ppa:kubuntu-ppa/backports
+language: cpp
+os:
+  - linux
+
+before_install:
+  - sudo sed -i 's/trusty/wily/g' /etc/apt/sources.list
   - sudo apt-get update -qq
-  - sudo apt-get -y -q -o Dpkg::Options::=--force-confdef -o Dpkg::Options::=--force-confnew dist-upgrade
-  - sudo apt-get install -y -q cmake extra-cmake-modules g++ git libkf5i18n-dev libkf5notifications-dev libkf5service-dev libkf5windowsystem-dev plasma-framework-dev qtbase5-dev qtdeclarative5-dev
+  - sudo apt-get upgrade
+
+addons:
+  apt:
+    sources:
+      - kubuntu-backports
+    packages:
+      - cmake
+      - extra-cmake-modules
+      - g++
+      - git
+      - libkf5i18n-dev
+      - libkf5notifications-dev
+      - libkf5service-dev
+      - libkf5windowsystem-dev
+      - plasma-framework-dev
+      - qtbase5-dev
+      - qtdeclarative5-dev
+
+before_script:
   - rm -rf build
   - mkdir build
