+++ sources/awesome-widget/plugin/awkeycache.h
@@ -1,0 +1,33 @@
+/***************************************************************************
+ *   This file is part of awesome-widgets                                  *
+ *                                                                         *
+ *   awesome-widgets is free software: you can redistribute it and/or      *
+ *   modify it under the terms of the GNU General Public License as        *
+ *   published by the Free Software Foundation, either version 3 of the    *
+ *   License, or (at your option) any later version.                       *
+ *                                                                         *
+ *   awesome-widgets is distributed in the hope that it will be useful,    *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with awesome-widgets. If not, see http://www.gnu.org/licenses/  *
+ ***************************************************************************/
+
+
+#ifndef AWKEYCACHE_H
+#define AWKEYCACHE_H
+
+#include <QHash>
+#include <QString>
+
+
+namespace AWKeyCache
+{
+void addKeyToCache(const QString type, const QString key = QString(""));
+QHash<QString, QStringList> loadKeysFromCache();
+};
+
+
+#endif /* AWKEYCACHE_H */
